Ignore contents of src/files but keep the directory
Harmony writes its output to the src/files directory. If the directort
is not present, outfiles are not created. The path to src/files is
hardcoded in element.cpp. The folder was added to the gitignore list to
avoid accidental commits of ouput files. The folder is kept in the repo
via .gitkeep