Create an installation folder from CMakeLists.txt
Related to issue #18

After a successful compilation, the user can execute "cmake install ." from the directory containing the VS Solution to create an installation folder in install/Harmony. This folder will contain the Harmony executable and all of the dlls it depends on, as well as the LICENSE file and the user guide (PDF). The Visual Studio Installer Project will use the install/Harmony folder retrieve the files for the MSI.